Neovim: More lazy, specially dependencies
Apparently adding any dependency on lazy, gives up any lazyness, the dep
will be loaded just before the dependant, regardless of whether it's
used or not.
Why? Don't know, it's dumb, folke on developers.md says u shouldn't use
dependencies when possible, but goes ahead and suggests it on other of
it's own plugins...
Was it that hard to add an option to deps to specify when to load?